Move assert to server-services-client
