+++ server/routerlicious/packages/common-utils/src/index.ts
@@ -20,4 +20,3 @@ export { toUtf8, fromBase64ToUtf8, fromUtf8ToBase64 } from "./base64Encoding";
 export { Buffer, IsoBuffer, bufferToString, stringToBuffer, Uint8ArrayToString } from "./buffer";
 export { gitHashFile } from "./hashFile";
 export { unreachableCase } from "./unreachable";
-export { assert } from "./assert";
