+++ server/routerlicious/packages/services-client/src/gitManager.ts
@@ -6,7 +6,7 @@
 import * as resources from "@fluidframework/gitresources";
 import { buildGitTreeHierarchy } from "@fluidframework/protocol-base";
 import * as api from "@fluidframework/protocol-definitions";
-import { assert } from "@fluidframework/server-common-utils";
+import { assert } from "./assert";
 import { debug } from "./debug";
 import {
 	ICreateRefParamsExternal,
